Trimming library utils
