+++ commands/auth.ts
@@ -42,7 +42,6 @@ const {
   getAccountId,
   addTestingOptions,
 } = require('../lib/commonOpts');
-const { logDebugInfo } = require('../lib/debugInfo');
 const { trackAuthAction, trackCommandUsage } = require('../lib/usageTracking');
 const { authenticateWithOauth } = require('../lib/oauth');
 const { EXIT_CODES } = require('../lib/enums/exitCodes');
@@ -75,7 +74,6 @@ exports.handler = async options => {
   const authType =
     (type && type.toLowerCase()) || PERSONAL_ACCESS_KEY_AUTH_METHOD.value;
   setLogLevel(options);
-  logDebugInfo(options);
 
   if (!getConfigPath(c)) {
     logger.error(i18n(`${i18nKey}.errors.noConfigFileFound`));
