+++ commands/create.ts
@@ -27,7 +27,6 @@ const fs = require('fs-extra');
 const { logError } = require('../lib/errorHandlers/index');
 const { logger } = require('@hubspot/local-dev-lib/logger');
 const { setLogLevel, getAccountId } = require('../lib/commonOpts');
-const { logDebugInfo } = require('../lib/debugInfo');
 const { resolveLocalPath } = require('../lib/filesystem');
 const { trackCommandUsage } = require('../lib/usageTracking');
 const assets = require('./create/index');
@@ -49,7 +48,6 @@ exports.handler = async options => {
   const { name, internal: getInternalVersion } = options;
 
   setLogLevel(options);
-  logDebugInfo(options);
   assetType = typeof assetType === 'string' && assetType.toLowerCase();
 
   if (assetType === 'global-partial') {
