+++ commands/feedback.ts
@@ -2,15 +2,20 @@
 const open = require('open');
 
 const { i18n } = require('../lib/lang');
-const { FEEDBACK_URLS } = require('../lib/constants');
 const { logger } = require('@hubspot/local-dev-lib/logger');
-
-const {
-  feedbackTypePrompt,
-  shouldOpenBrowserPrompt,
-} = require('../lib/prompts/feedbackPrompt');
+const { confirmPrompt, listPrompt } = require('../lib/prompts/promptUtils');
 
 const i18nKey = 'commands.project.subcommands.feedback';
+
+const FEEDBACK_OPTIONS = {
+  BUG: 'bug',
+  GENERAL: 'general',
+};
+const FEEDBACK_URLS = {
+  BUG: 'https://github.com/HubSpot/hubspot-cli/issues/new',
+  GENERAL:
+    'https://docs.google.com/forms/d/e/1FAIpQLSejZZewYzuH3oKBU01tseX-cSWOUsTHLTr-YsiMGpzwcvgIMg/viewform?usp=sf_link',
+};
 
 exports.command = 'feedback';
 exports.describe = i18n(`${i18nKey}.describe`);
@@ -19,8 +24,16 @@ exports.handler = async options => {
   const { bug: bugFlag, general: generalFlag } = options;
   const usedTypeFlag = bugFlag !== generalFlag;
 
-  const { type } = await feedbackTypePrompt(usedTypeFlag);
-  const { shouldOpen } = await shouldOpenBrowserPrompt(type, usedTypeFlag);
+  await listPrompt(i18n(`${i18nKey}.feedbackType.prompt`), {
+    choices: Object.values(FEEDBACK_OPTIONS).map(option => ({
+      name: i18n(`${i18nKey}.feedbackType.${option}`),
+      value: option,
+    })),
+    when: !usedTypeFlag,
+  });
+  const shouldOpen = await confirmPrompt(i18n(`${i18nKey}.openPrompt`), {
+    when: !usedTypeFlag,
+  });
 
   if (shouldOpen || usedTypeFlag) {
     // NOTE: for now, all feedback should go to the hubspot-cli repository
