+++ commands/functions/list.ts
@@ -1,8 +1,8 @@
 // @ts-nocheck
+const moment = require('moment');
 const { getRoutes } = require('@hubspot/local-dev-lib/api/functions');
 const { logger } = require('@hubspot/local-dev-lib/logger');
 const { logError, ApiErrorContext } = require('../../lib/errorHandlers/index');
-const { getFunctionArrays } = require('../../lib/getFunctionArrays');
 const { getTableContents, getTableHeader } = require('../../lib/ui/table');
 const {
   addConfigOptions,
@@ -19,6 +19,19 @@ const { EXIT_CODES } = require('../../lib/enums/exitCodes');
 
 exports.command = 'list';
 exports.describe = i18n(`${i18nKey}.describe`);
+
+const getFunctionArrays = resp => {
+  return resp.objects.map(func => {
+    const { route, method, created, updated, secretNames } = func;
+    return [
+      route,
+      method,
+      secretNames.join(', '),
+      moment(created).format(),
+      moment(updated).format(),
+    ];
+  });
+};
 
 exports.handler = async options => {
   loadAndValidateOptions(options);
