+++ commands/init.ts
@@ -39,7 +39,6 @@ const {
   personalAccessKeyPrompt,
 } = require('../lib/prompts/personalAccessKeyPrompt');
 const { cliAccountNamePrompt } = require('../lib/prompts/accountNamePrompt');
-const { logDebugInfo } = require('../lib/debugInfo');
 const { authenticateWithOauth } = require('../lib/oauth');
 const { EXIT_CODES } = require('../lib/enums/exitCodes');
 const { uiFeatureHighlight } = require('../lib/ui');
@@ -108,7 +107,6 @@ exports.handler = async options => {
   } = options;
   const configPath = (c && path.join(getCwd(), c)) || getConfigPath();
   setLogLevel(options);
-  logDebugInfo(options);
   trackCommandUsage('init', {
     authType,
   });
