+++ commands/project/__tests__/logs.test.ts
@@ -4,7 +4,7 @@ const {
   addUseEnvironmentOptions,
   getAccountId,
 } = require('../../../lib/commonOpts');
-const ProjectLogsManager = require('../../../lib/projectLogsManager');
+const ProjectLogsManager = require('../../../lib/ProjectLogsManager');
 const {
   projectLogsPrompt,
 } = require('../../../lib/prompts/projectsLogsPrompt');
@@ -19,7 +19,7 @@ jest.mock('@hubspot/local-dev-lib/logger');
 jest.mock('../../../lib/commonOpts');
 jest.mock('../../../lib/usageTracking');
 jest.mock('../../../lib/validation');
-jest.mock('../../../lib/projectLogsManager');
+jest.mock('../../../lib/ProjectLogsManager');
 jest.mock('../../../lib/prompts/projectsLogsPrompt');
 jest.mock('../../../lib/ui/table');
 jest.mock('../../../lib/errorHandlers');
