+++ commands/project/logs.ts
@@ -18,7 +18,7 @@ const { uiBetaTag, uiLine, uiLink } = require('../../lib/ui');
 const { projectLogsPrompt } = require('../../lib/prompts/projectsLogsPrompt');
 const { i18n } = require('../../lib/lang');
 const { EXIT_CODES } = require('../../lib/enums/exitCodes');
-const ProjectLogsManager = require('../../lib/projectLogsManager');
+const ProjectLogsManager = require('../../lib/ProjectLogsManager');
 
 const i18nKey = 'commands.project.subcommands.logs';
