+++ commands/sandbox/delete.ts
@@ -20,9 +20,7 @@ const {
   removeSandboxAccountFromConfig,
   updateDefaultAccount,
 } = require('@hubspot/local-dev-lib/config');
-const {
-  selectAndSetAsDefaultAccountPrompt,
-} = require('../../lib/prompts/accountsPrompt');
+const { selectAccountFromConfig } = require('../../lib/prompts/accountsPrompt');
 const { EXIT_CODES } = require('../../lib/enums/exitCodes');
 const { promptUser } = require('../../lib/prompts/promptUtils');
 const { getHubSpotWebsiteOrigin } = require('@hubspot/local-dev-lib/urls');
@@ -157,7 +155,8 @@ exports.handler = async options => {
       sandboxAccountId
     );
     if (promptDefaultAccount && !force) {
-      await selectAndSetAsDefaultAccountPrompt(getConfig());
+      const newDefaultAccount = await selectAccountFromConfig(getConfig());
+      updateDefaultAccount(newDefaultAccount);
     } else {
       // If force is specified, skip prompt and set the parent account id as the default account
       updateDefaultAccount(parentAccountId);
@@ -209,7 +208,8 @@ exports.handler = async options => {
         sandboxAccountId
       );
       if (promptDefaultAccount && !force) {
-        await selectAndSetAsDefaultAccountPrompt(getConfig());
+        const newDefaultAccount = await selectAccountFromConfig(getConfig());
+        updateDefaultAccount(newDefaultAccount);
       } else {
         // If force is specified, skip prompt and set the parent account id as the default account
         updateDefaultAccount(parentAccountId);
