+++ commands/theme/marketplace-validate.ts
@@ -14,7 +14,7 @@ const {
   fetchValidationResults,
   processValidationErrors,
   displayValidationResults,
-} = require('../../lib/marketplace-validate');
+} = require('../../lib/marketplaceValidate');
 const { i18n } = require('../../lib/lang');
 
 const i18nKey = 'commands.theme.subcommands.marketplaceValidate';
