+++ lib/ProjectLogsManager.ts
