+++ lib/generateSelectors.ts
