--- lib/getFunctionArrays.ts
@@ -1,19 +1,0 @@
-// @ts-nocheck
-const moment = require('moment');
-
-const getFunctionArrays = resp => {
-  return resp.objects.map(func => {
-    const { route, method, created, updated, secretNames } = func;
-    return [
-      route,
-      method,
-      secretNames.join(', '),
-      moment(created).format(),
-      moment(updated).format(),
-    ];
-  });
-};
-
-module.exports = {
-  getFunctionArrays,
-};
