+++ lib/marketplaceValidate.ts
