+++ lib/prompts/accountsPrompt.ts
@@ -1,5 +1,4 @@
 // @ts-nocheck
-const { updateDefaultAccount } = require('@hubspot/local-dev-lib/config');
 const { promptUser } = require('./promptUtils');
 const { i18n } = require('../lang');
 const { uiAccountDescription } = require('../ui');
@@ -28,23 +27,6 @@ const selectAccountFromConfig = async (config, prompt) => {
   return selectedDefault;
 };
 
-const selectAndSetAsDefaultAccountPrompt = async config => {
-  const { default: selectedDefault } = await promptUser([
-    {
-      type: 'list',
-      look: false,
-      name: 'default',
-      pageSize: 20,
-      message: i18n(`${i18nKey}.promptMessage`),
-      choices: mapAccountChoices(config.portals),
-      default: config.defaultPortal,
-    },
-  ]);
-  updateDefaultAccount(selectedDefault);
-};
-
 module.exports = {
-  selectAndSetAsDefaultAccountPrompt,
   selectAccountFromConfig,
-  mapAccountChoices,
 };
