+++ lib/prompts/promptUtils.ts
@@ -1,24 +1,45 @@
-// @ts-nocheck
 const inquirer = require('inquirer');
 
-const promptUser = async promptConfig => {
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export const promptUser = async (promptConfig: any) => {
   const prompt = inquirer.createPromptModule();
   return prompt(promptConfig);
 };
 
-const confirmPrompt = async (message, defaultAnswer = true) => {
+export const confirmPrompt = async (
+  message: string,
+  {
+    defaultAnswer = true,
+    when,
+  }: { defaultAnswer: boolean; when?: boolean | (() => boolean) }
+): Promise<boolean> => {
   const { choice } = await promptUser([
     {
       name: 'choice',
       type: 'confirm',
-      default: defaultAnswer,
       message,
+      default: defaultAnswer,
+      when,
     },
   ]);
   return choice;
 };
 
-module.exports = {
-  promptUser,
-  confirmPrompt,
+export const listPrompt = async (
+  message: string,
+  {
+    choices,
+    when,
+  }: { choices: { [key: string]: string }[]; when?: boolean | (() => boolean) }
+): Promise<boolean> => {
+  const { choice } = await promptUser([
+    {
+      name: 'choice',
+      type: 'list',
+      message,
+      choices,
+      when,
+    },
+  ]);
+  return choice;
 };
