--- lib/regex.ts
@@ -1,6 +1,0 @@
-// @ts-nocheck
-const STRING_WITH_NO_SPACES_REGEX = /^\S*$/;
-
-module.exports = {
-  STRING_WITH_NO_SPACES_REGEX,
-};
